refactor(sdk-lib-mpc): move ed25519 hd tree
TICKET: WP-400
