+++ modules/account-lib/src/index.ts
@@ -2,8 +2,8 @@ import {
   BaseBuilder,
   BaseTransactionBuilderFactory,
   BuildTransactionError,
-  Ed25519BIP32,
   Ed25519KeyDeriver,
+  Ed25519BIP32,
   Eddsa,
   accountLibBaseCoin,
   acountLibCrypto,
