+++ modules/sdk-coin-sol/package.json
@@ -41,6 +41,7 @@
   },
   "dependencies": {
     "@bitgo/sdk-core": "^8.15.0",
+    "@bitgo/sdk-lib-mpc": "^8.5.0",
     "@bitgo/statics": "^18.4.0",
     "@solana/spl-token": "0.3.1",
     "@solana/web3.js": "1.78.0",
