+++ modules/sdk-coin-sui/package.json
@@ -42,6 +42,7 @@
   "dependencies": {
     "@bitgo/blake2b": "^3.2.4",
     "@bitgo/sdk-core": "^8.15.0",
+    "@bitgo/sdk-lib-mpc": "^8.5.0",
     "@bitgo/statics": "^18.4.0",
     "@mysten/bcs": "^0.7.0",
     "bignumber.js": "^9.0.0",
