+++ modules/sdk-lib-mpc/src/curves/constant.ts
@@ -1,0 +1,2 @@
+// 2^256
+export const chaincodeBase = BigInt('0x010000000000000000000000000000000000000000000000000000000000000000');
