+++ modules/sdk-lib-mpc/src/curves/index.ts
@@ -1,5 +1,7 @@
 export * from './types';
 export * from './util';
-export * from './bip32HdTree';
+export * from './constant';
 export * from './secp256k1';
+export * from './secp256k1Bip32HdTree';
 export * from './ed25519';
+export * from './ed25519Bip32HdTree';
